Update deployment script to v1.2.2

+++ .config/azure/deploy.sh
@@ -2,7 +2,7 @@
 
 # ----------------------
 # Meteor Azure
-# Version: 1.2.0
+# Version: 1.2.2
 # ----------------------
 
 # ----------------------
@@ -41,6 +41,29 @@ fi
 # Prerequisites
 # ------------
 
+# Validate configuration
+if [ -e "$DEPLOYMENT_SOURCE\iisnode.yml" ]; then
+  echo "meteor-azure: WARNING! iisnode.yml will not be respected, please move configuration to web.config"
+fi
+if [ ! -e "$DEPLOYMENT_SOURCE\.config\azure\web.config" ]; then
+  echo "meteor-azure: WARNING! No web.config was found (app may not start properly)"
+fi
+if [ $SCM_COMMAND_IDLE_TIMEOUT != 7200 ]; then
+  echo "meteor-azure: WARNING! Not using recommended 'SCM_COMMAND_IDLE_TIMEOUT' (build may abort unexpectedly)"
+fi
+if [[ ! -v METEOR_AZURE_NODE_VERSION ]]; then
+  echo "meteor-azure: ERROR! You must specify App Setting 'METEOR_AZURE_NODE_VERSION'"
+  exit 1
+fi
+if [[ ! -v METEOR_AZURE_NPM_VERSION ]]; then
+  echo "meteor-azure: ERROR! You must specify App Setting 'METEOR_AZURE_NPM_VERSION'"
+  exit 1
+fi
+if [[ ! -v ROOT_URL ]]; then
+  echo "meteor-azure: ERROR! You must specify App Setting 'ROOT_URL'"
+  exit 1
+fi
+
 # Prepare installation scope
 if [ ! -d D:/home/meteor-azure ];
   then mkdir D:/home/meteor-azure
@@ -68,7 +91,9 @@ fi
 echo meteor-azure: Setting Node version
 export NVM_HOME=D:/home/meteor-azure/nvm
 nvm/nvm.exe install $METEOR_AZURE_NODE_VERSION 32
-cp "nvm/v$METEOR_AZURE_NODE_VERSION/node32.exe" "nvm/v$METEOR_AZURE_NODE_VERSION/node.exe"
+if [ ! -e "nvm/v$METEOR_AZURE_NODE_VERSION/node.exe" ]; then
+  cp "nvm/v$METEOR_AZURE_NODE_VERSION/node32.exe" "nvm/v$METEOR_AZURE_NODE_VERSION/node.exe"
+fi
 export PATH="$HOME/meteor-azure/nvm/v$METEOR_AZURE_NODE_VERSION:$PATH"
 echo "meteor-azure: Now using Node $(node -v) (32-bit)"
